Reorganize collector/solidifier structure. Makr Syncer wait to spawn until a message is received on the milestone stream. Request from INX directly in syncer. Spawn solidifiers under syncer for that process.
