+++ bin/inx-chronicle/src/collector/mod.rs
@@ -6,17 +6,21 @@ pub mod solidifier;
 #[cfg(all(feature = "stardust", feature = "inx"))]
 pub(crate) mod stardust_inx;
 
-use std::collections::HashMap;
-
 use async_trait::async_trait;
 use chronicle::{
-    db::{bson::DocError, MongoDb},
-    runtime::{Actor, ActorContext, ActorError, Addr, ConfigureActor, HandleEvent, Report, RuntimeError},
+    db::{
+        bson::DocError,
+        model::stardust::{message::MessageRecord, milestone::MilestoneRecord},
+        MongoDb,
+    },
+    runtime::{Actor, ActorContext, ActorError, HandleEvent, Report, RuntimeError},
+    types::ledger::Metadata,
 };
-pub use config::CollectorConfig;
 use mongodb::bson::document::ValueAccessError;
-use solidifier::Solidifier;
 use thiserror::Error;
+
+pub use self::config::CollectorConfig;
+use self::solidifier::Solidifier;
 
 #[derive(Debug, Error)]
 pub enum CollectorError {
@@ -42,28 +46,17 @@ impl Collector {
     }
 }
 
-pub struct CollectorState {
-    solidifiers: HashMap<usize, Addr<Solidifier>>,
-}
-
 #[async_trait]
 impl Actor for Collector {
-    type State = CollectorState;
+    type State = ();
     type Error = CollectorError;
 
     async fn init(&mut self, cx: &mut ActorContext<Self>) -> Result<Self::State, Self::Error> {
-        let mut solidifiers = HashMap::new();
-        for i in 0..self.config.solidifier_count {
-            solidifiers.insert(
-                i,
-                cx.spawn_child(Solidifier::new(i, self.db.clone()).with_registration(false))
-                    .await,
-            );
-        }
+        cx.spawn_child(Solidifier::new(0, self.db.clone())).await;
         #[cfg(all(feature = "stardust", feature = "inx"))]
         cx.spawn_child(stardust_inx::InxWorker::new(self.db.clone(), self.config.inx.clone()))
             .await;
-        Ok(CollectorState { solidifiers })
+        Ok(())
     }
 }
 
@@ -73,7 +66,7 @@ impl HandleEvent<Report<Solidifier>> for Collector {
         &mut self,
         cx: &mut ActorContext<Self>,
         event: Report<Solidifier>,
-        state: &mut Self::State,
+        _state: &mut Self::State,
     ) -> Result<(), Self::Error> {
         match event {
             Report::Success(_) => {
@@ -83,9 +76,7 @@ impl HandleEvent<Report<Solidifier>> for Collector {
                 ActorError::Result(e) => match e {
                     #[cfg(all(feature = "stardust", feature = "inx"))]
                     solidifier::SolidifierError::MissingStardustInxRequester => {
-                        state
-                            .solidifiers
-                            .insert(report.actor.id, cx.spawn_child(report.actor).await);
+                        cx.spawn_child(report.actor).await;
                     }
                     // TODO: Maybe map Solidifier errors to Collector errors and return them?
                     _ => {
@@ -98,5 +89,213 @@ impl HandleEvent<Report<Solidifier>> for Collector {
             },
         }
         Ok(())
+    }
+}
+
+#[cfg(all(feature = "stardust", feature = "inx"))]
+mod _stardust_inx {
+    use super::{
+        stardust_inx::{InxWorker, InxWorkerError, RequestedMessage},
+        *,
+    };
+    use crate::collector::stardust_inx::MilestoneState;
+
+    #[async_trait]
+    impl HandleEvent<Report<InxWorker>> for Collector {
+        async fn handle_event(
+            &mut self,
+            cx: &mut ActorContext<Self>,
+            event: Report<InxWorker>,
+            _state: &mut Self::State,
+        ) -> Result<(), Self::Error> {
+            match event {
+                Report::Success(_) => {
+                    cx.shutdown();
+                }
+                Report::Error(report) => match report.error {
+                    ActorError::Result(e) => match e {
+                        InxWorkerError::ConnectionError(_) => {
+                            let wait_interval = self.config.inx.connection_retry_interval;
+                            log::info!("Retrying INX connection in {} seconds.", wait_interval.as_secs_f32());
+                            cx.delay(
+                                chronicle::runtime::SpawnActor::new(InxWorker::new(
+                                    self.db.clone(),
+                                    self.config.inx.clone(),
+                                )),
+                                wait_interval,
+                            )?;
+                        }
+                        // TODO: This is stupid, but we can't use the ErrorKind enum so :shrug:
+                        InxWorkerError::TransportFailed(e) => match e.to_string().as_ref() {
+                            "transport error" => {
+                                cx.spawn_child(InxWorker::new(self.db.clone(), self.config.inx.clone()))
+                                    .await;
+                            }
+                            _ => {
+                                cx.shutdown();
+                            }
+                        },
+                        InxWorkerError::MissingCollector => {
+                            cx.delay(
+                                chronicle::runtime::SpawnActor::new(InxWorker::new(
+                                    self.db.clone(),
+                                    self.config.inx.clone(),
+                                )),
+                                None,
+                            )?;
+                        }
+                        InxWorkerError::MongoDb(e) => {
+                            return Err(CollectorError::MongoDb(e));
+                        }
+                        InxWorkerError::ListenerError(_)
+                        | InxWorkerError::Runtime(_)
+                        | InxWorkerError::Read(_)
+                        | InxWorkerError::ParsingAddressFailed(_)
+                        | InxWorkerError::InvalidAddress(_)
+                        | InxWorkerError::FailedToAnswerRequest
+                        | InxWorkerError::InxTypeConversion(_) => {
+                            cx.shutdown();
+                        }
+                    },
+                    ActorError::Panic | ActorError::Aborted => {
+                        cx.shutdown();
+                    }
+                },
+            }
+            Ok(())
+        }
+    }
+
+    #[async_trait]
+    impl HandleEvent<inx::proto::Message> for Collector {
+        async fn handle_event(
+            &mut self,
+            _cx: &mut ActorContext<Self>,
+            message: inx::proto::Message,
+            _state: &mut Self::State,
+        ) -> Result<(), Self::Error> {
+            log::trace!(
+                "Received Stardust Message Event ({})",
+                prefix_hex::encode(message.message_id.as_ref().unwrap().id.as_slice())
+            );
+            match MessageRecord::try_from(message) {
+                Ok(rec) => {
+                    self.db.upsert_message_record(&rec).await?;
+                }
+                Err(e) => {
+                    log::error!("Could not read message: {:?}", e);
+                }
+            };
+            Ok(())
+        }
+    }
+
+    #[async_trait]
+    impl HandleEvent<inx::proto::MessageMetadata> for Collector {
+        async fn handle_event(
+            &mut self,
+            _cx: &mut ActorContext<Self>,
+            metadata: inx::proto::MessageMetadata,
+            _state: &mut Self::State,
+        ) -> Result<(), Self::Error> {
+            log::trace!(
+                "Received Stardust Message Referenced Event ({})",
+                metadata.milestone_index
+            );
+            match inx::MessageMetadata::try_from(metadata) {
+                Ok(rec) => {
+                    let message_id = rec.message_id;
+                    self.db
+                        .update_message_metadata(&message_id.into(), &Metadata::from(rec))
+                        .await?;
+                }
+                Err(e) => {
+                    log::error!("Could not read message metadata: {:?}", e);
+                }
+            };
+            Ok(())
+        }
+    }
+
+    #[async_trait]
+    impl HandleEvent<inx::proto::Milestone> for Collector {
+        async fn handle_event(
+            &mut self,
+            cx: &mut ActorContext<Self>,
+            milestone: inx::proto::Milestone,
+            _state: &mut Self::State,
+        ) -> Result<(), Self::Error> {
+            log::trace!(
+                "Received Stardust Milestone Event ({})",
+                milestone.milestone_info.as_ref().unwrap().milestone_index
+            );
+            match MilestoneRecord::try_from(milestone) {
+                Ok(rec) => {
+                    self.db.upsert_milestone_record(&rec).await?;
+                    // Get or create the milestone state
+                    let mut ms_state = MilestoneState::new(rec.milestone_index);
+                    ms_state
+                        .process_queue
+                        .extend(Vec::from(rec.payload.essence.parents).into_iter());
+                    cx.addr::<Solidifier>().await.send(ms_state)?;
+                }
+                Err(e) => {
+                    log::error!("Could not read milestone: {:?}", e);
+                }
+            }
+            Ok(())
+        }
+    }
+
+    #[async_trait]
+    impl<Sender: Actor> HandleEvent<RequestedMessage<Sender>> for Collector
+    where
+        Sender: 'static + HandleEvent<MilestoneState>,
+    {
+        async fn handle_event(
+            &mut self,
+            _cx: &mut ActorContext<Self>,
+            RequestedMessage {
+                raw,
+                metadata,
+                sender_addr,
+                ms_state,
+            }: RequestedMessage<Sender>,
+            _state: &mut Self::State,
+        ) -> Result<(), Self::Error> {
+            match raw {
+                Some(raw) => {
+                    log::trace!("Received Stardust Requested Message and Metadata");
+                    match MessageRecord::try_from((raw, metadata)) {
+                        Ok(rec) => {
+                            self.db.upsert_message_record(&rec).await?;
+                            // Send this directly to the solidifier that requested it
+                            sender_addr.send(ms_state)?;
+                        }
+                        Err(e) => {
+                            log::error!("Could not read message: {:?}", e);
+                        }
+                    };
+                }
+                None => {
+                    log::trace!("Received Stardust Requested Metadata");
+                    match inx::MessageMetadata::try_from(metadata) {
+                        Ok(rec) => {
+                            let message_id = rec.message_id;
+                            self.db
+                                .update_message_metadata(&message_id.into(), &Metadata::from(rec))
+                                .await?;
+                            // Send this directly to the solidifier that requested it
+                            sender_addr.send(ms_state)?;
+                        }
+                        Err(e) => {
+                            log::error!("Could not read message metadata: {:?}", e);
+                        }
+                    };
+                }
+            }
+
+            Ok(())
+        }
     }
 }
