+++ bin/inx-chronicle/src/collector/solidifier.rs
@@ -4,7 +4,7 @@
 use async_trait::async_trait;
 use chronicle::{
     db::MongoDb,
-    runtime::{Actor, ActorContext, RuntimeError},
+    runtime::{Actor, ActorContext, HandleEvent, RuntimeError},
 };
 use mongodb::bson::document::ValueAccessError;
 use thiserror::Error;
@@ -45,5 +45,93 @@ impl Actor for Solidifier {
 
     fn name(&self) -> std::borrow::Cow<'static, str> {
         format!("Solidifier {}", self.id).into()
+    }
+}
+
+#[cfg(all(feature = "stardust", feature = "inx"))]
+mod _stardust_inx {
+    use super::*;
+    use crate::collector::stardust_inx::{InxWorker, MessageRequest, MetadataRequest, MilestoneState};
+
+    #[async_trait]
+    impl HandleEvent<MilestoneState> for Solidifier {
+        async fn handle_event(
+            &mut self,
+            cx: &mut ActorContext<Self>,
+            mut ms_state: MilestoneState,
+            _state: &mut Self::State,
+        ) -> Result<(), Self::Error> {
+            // Process by iterating the queue until we either complete the milestone or fail to find a message
+            while let Some(message_id) = ms_state.process_queue.front() {
+                // First check if we already processed this message in this run
+                if ms_state.visited.contains(message_id) {
+                    ms_state.process_queue.pop_front();
+                } else {
+                    // Try the database
+                    match self.db.get_message(message_id).await? {
+                        Some(message_rec) => {
+                            match message_rec
+                                .metadata
+                                .map(|metadata| metadata.referenced_by_milestone_index)
+                            {
+                                Some(ms_index) => {
+                                    log::trace!(
+                                        "Message {} is referenced by milestone {}",
+                                        message_id.to_hex(),
+                                        ms_index
+                                    );
+
+                                    // We add the current message to the list of visited messages.
+                                    ms_state.visited.insert(message_id.clone());
+
+                                    // We may have reached a different milestone, in which case there is nothing to
+                                    // do for this message
+                                    if ms_state.milestone_index == ms_index {
+                                        let parents = Vec::from(message_rec.message.parents);
+                                        ms_state.process_queue.extend(parents);
+                                    }
+                                    ms_state.process_queue.pop_front();
+                                }
+                                // If the message has not been referenced, we can't proceed
+                                None => {
+                                    log::trace!("Requesting metadata for message {}", message_id.to_hex());
+                                    // Send the state and everything. If the requester finds the message, it will circle
+                                    // back.
+                                    cx.addr::<InxWorker>()
+                                        .await
+                                        .send(MetadataRequest::new(message_id.clone(), cx.handle().clone(), ms_state))
+                                        .map_err(|_| SolidifierError::MissingStardustInxRequester)?;
+                                    return Ok(());
+                                }
+                            }
+                        }
+                        // Otherwise, send a message to the requester
+                        None => {
+                            log::trace!("Requesting message {}", message_id.to_hex());
+                            // Send the state and everything. If the requester finds the message, it will circle
+                            // back.
+                            cx.addr::<InxWorker>()
+                                .await
+                                .send(MessageRequest::new(message_id.clone(), cx.handle().clone(), ms_state))
+                                .map_err(|_| SolidifierError::MissingStardustInxRequester)?;
+                            return Ok(());
+                        }
+                    }
+                }
+            }
+
+            // If we finished all the parents, that means we have a complete milestone
+            // so we should mark it synced
+            self.db.upsert_sync_record(ms_state.milestone_index).await?;
+
+            log::info!("Milestone {} solidified.", ms_state.milestone_index,);
+
+            // If this was sent with a channel, respond to it
+            if let Some(tx) = ms_state.sender {
+                tx.send(ms_state.milestone_index).ok();
+            }
+
+            Ok(())
+        }
     }
 }
